+++ xclim/indicators/atmos/_temperature.py
@@ -54,6 +54,8 @@ __all__ = [
     "growing_season_end",
     "tropical_nights",
     "degree_days_exceedance_date",
+    "warm_spell_duration_index",
+    "maximum_consecutive_warm_days",
 ]
 
 
@@ -644,4 +646,24 @@ degree_days_exceedance_date = Tas(
     " exceeds {sum_thresh}, the cumulative sum starts on {after_date}.",
     cell_methods="",
     compute=indices.degree_days_exceedance_date,
+)
+
+
+warm_spell_duration_index = Tasmax(
+    identifier="warm_spell_duration_index",
+    description="{freq} total number of days within spells of at least {window} days with tmax above the 90th daily percentile.",
+    units="days",
+    standard_name="number_of_days_with_air_temperature_above_threshold",
+    cell_methods="time: sum over days",
+    compute=indices.warm_spell_duration_index,
+)
+
+
+maximum_consecutive_warm_days = Tasmax(
+    identifier="maximum_consecutive_warm_days",
+    description="{freq} longest spell of consecutive days with Tmax above {thresh}.",
+    units="days",
+    standard_name="spell_length_of_days_with_air_temperature_above_threshold",
+    cell_methods="time: maximum over days",
+    compute=indices.maximum_consecutive_tx_days,
 )