+++ xclim/indices/_anuclim.py
@@ -70,11 +70,11 @@ def isothermality(
     Parameters
     ----------
     tasmin : xarray.DataArray
-      Average daily minimum temperature [℃] or [K] at daily, weekly, or monthly frequency.
+      Average daily minimum temperature at daily, weekly, or monthly frequency.
     tasmax : xarray.DataArray
-      Average daily maximum temperature [℃] or [K] at daily, weekly, or monthly frequency.
+      Average daily maximum temperature at daily, weekly, or monthly frequency.
     freq : str
-      Resampling frequency; Defaults to "YS".
+      Resampling frequency.
 
     Returns
     -------
@@ -106,7 +106,7 @@ def temperature_seasonality(tas: xarray.DataArray) -> xarray.DataArray:
     Parameters
     ----------
     tas : xarray.DataArray
-      Mean temperature [℃] or [K] at daily, weekly, or monthly frequency.
+      Mean temperature at daily, weekly, or monthly frequency.
 
     Returns
     -------
@@ -212,13 +212,13 @@ def tg_mean_warmcold_quarter(
     Parameters
     ----------
     tas : xarray.DataArray
-      Mean temperature [℃] or [K] at daily, weekly, or monthly frequency.
-    op : str
+      Mean temperature at daily, weekly, or monthly frequency.
+    op : str {'warmest', 'coldest'}
       Operation to perform:  'warmest' calculate warmest quarter; 'coldest' calculate coldest quarter.
     src_timestep : {'D', 'W', 'M'}
       Input data time frequency - One of daily, weekly or monthly.
     freq : str
-      Resampling frequency; Defaults to "YS".
+      Resampling frequency.
 
     Returns
     -------
@@ -265,15 +265,15 @@ def tg_mean_wetdry_quarter(
     Parameters
     ----------
     tas : xarray.DataArray
-      Mean temperature [℃] or [K] at daily, weekly, or monthly frequency.
+      Mean temperature at daily, weekly, or monthly frequency.
     pr : xarray.DataArray
       Total precipitation rate at daily, weekly, or monthly frequency.
-    op : str {'wettest', 'driest'}
+    op : {'wettest', 'driest'}
       Operation to perform: 'wettest' calculate for the wettest quarter; 'driest' calculate for the driest quarter.
     src_timestep : {'D', 'W', 'M'}
       Input data time frequency - One of daily, weekly or monthly.
     freq : str
-      Resampling frequency; Defaults to "YS".
+      Resampling frequency.
 
     Returns
     -------
@@ -311,12 +311,12 @@ def prcptot_wetdry_quarter(
     ----------
     pr : xarray.DataArray
       Total precipitation rate at daily, weekly, or monthly frequency.
-    op : str
+    op : {'wettest', 'driest'}
       Operation to perform :  'wettest' calculate wettest quarter ; 'driest' calculate driest quarter.
     src_timestep : {'D', 'W', 'M'}
       Input data time frequency - One of daily, weekly or monthly.
     freq : str
-      Resampling frequency; Defaults to "YS".
+      Resampling frequency.
 
     Returns
     -------
@@ -372,13 +372,13 @@ def prcptot_warmcold_quarter(
     pr : xarray.DataArray
       Total precipitation rate at daily, weekly, or monthly frequency.
     tas : xarray.DataArray
-      Mean temperature [℃] or [K] at daily, weekly, or monthly frequency.
-    op : str
+      Mean temperature at daily, weekly, or monthly frequency.
+    op : {'warmest', 'coldest'}
       Operation to perform: 'warmest' calculate for the warmest quarter ; 'coldest' calculate for the coldest quarter.
     src_timestep : {'D', 'W', 'M'}
       Input data time frequency - One of daily, weekly or monthly.
     freq : str
-      Resampling frequency; Defaults to "YS".
+      Resampling frequency.
 
     Returns
     -------
@@ -416,7 +416,7 @@ def prcptot(
     src_timestep : {'D', 'W', 'M'}
       Input data time frequency - One of daily, weekly or monthly.
     freq : str
-      Resampling frequency; Defaults to "YS".
+      Resampling frequency.
 
     Returns
     -------
@@ -443,12 +443,12 @@ def prcptot_wetdry_period(
     ----------
     pr : xarray.DataArray
       Total precipitation flux [mm d-1], [mm week-1], [mm month-1] or similar.
-    op : str
+    op : {'wettest', 'driest'}
       Operation to perform :  'wettest' calculate wettest period ; 'driest' calculate driest period.
     src_timestep : {'D', 'W', 'M'}
       Input data time frequency - One of daily, weekly or monthly.
     freq : str
-      Resampling frequency; Defaults to "YS".
+      Resampling frequency.
 
     Returns
     -------