+++ xclim/testing/tests/test_indicators.py
@@ -2,6 +2,7 @@
 # -*- coding: utf-8 -*-
 # Tests for the Indicator objects
 import gc
+import json
 from typing import Union
 
 import dask
@@ -18,9 +19,9 @@ from xclim.core.formatting import (
     parse_doc,
     update_history,
 )
-from xclim.core.indicator import Daily, Indicator, InputKind, registry
+from xclim.core.indicator import Daily, Indicator, registry
 from xclim.core.units import units
-from xclim.core.utils import MissingVariableError
+from xclim.core.utils import InputKind, MissingVariableError
 from xclim.indices import tg_mean
 from xclim.indices.generic import select_time
 from xclim.testing import open_dataset
@@ -109,7 +110,7 @@ def test_registering():
 
     # Confirm registries live in subclasses.
     class IndicatorNew(Indicator):
-        _nvar = 2
+        nvar = 2
 
     # Identifier must be given
     with pytest.raises(AttributeError, match="has not been set."):
@@ -239,6 +240,33 @@ def test_json(pr_series):
         assert set(output.keys()).issubset(output_exp)
 
 
+def test_all_jsonable(official_indicators):
+    problems = []
+    for identifier, ind in official_indicators.items():
+        indinst = ind.get_instance()
+        try:
+            json.dumps(indinst.json())
+        except TypeError:
+            problems.append(identifier)
+    if problems:
+        raise ValueError(
+            f"Indicators {problems} provide problematic json serialization."
+        )
+
+
+def test_all_parameters_understood(official_indicators):
+    problems = []
+    for identifier, ind in official_indicators.items():
+        indinst = ind.get_instance()
+        for name, param in indinst.parameters.items():
+            if param["kind"] == InputKind.OTHER_PARAMETER:
+                problems.append((identifier, name))
+    if problems:
+        raise ValueError(
+            f"The following indicator/parameter couple {problems} use types not listed in InputKind."
+        )
+
+
 def test_signature():
     from inspect import signature
 
@@ -254,7 +282,7 @@ def test_signature():
 
 def test_doc():
     ind = UniIndTemp()
-    assert ind.__call__.__doc__ == ind.compute.__doc__
+    assert ind.__call__.__doc__.startswith("Docstring (realm: atmos)")
 
 
 def test_delayed(tasmax_series):
@@ -300,13 +328,8 @@ def test_parse_doc():
         doc["abstract"]
         == "Resample the original daily mean temperature series by taking the mean over each period."
     )
-    assert (
-        doc["parameters"]["tas"]["description"] == "Mean daily temperature [℃] or [K]"
-    )
-    assert (
-        doc["parameters"]["freq"]["description"]
-        == 'Resampling frequency; Defaults to "YS" (yearly).'
-    )
+    assert doc["parameters"]["tas"]["description"] == "Mean daily temperature."
+    assert doc["parameters"]["freq"]["description"] == "Resampling frequency."
     assert doc["notes"].startswith("Let")
     assert "math::" in doc["notes"]
     assert "references" not in doc
@@ -326,13 +349,15 @@ def test_parsed_doc():
 
     params = xclim.atmos.drought_code.parameters
     assert params["tas"]["description"] == "Noon temperature."
-    assert params["tas"]["annotation"] is Union[str, xr.DataArray]
+    assert params["tas"]["units"] == "[temperature]"
     assert params["tas"]["kind"] is InputKind.VARIABLE
     assert params["tas"]["default"] == "tas"
     assert params["snd"]["default"] is None
     assert params["snd"]["kind"] is InputKind.OPTIONAL_VARIABLE
-    assert params["shut_down_mode"]["annotation"] is str
-    assert params["shut_down_mode"]["kind"] is InputKind.PARAMETER
+    assert params["snd"]["units"] == "[length]"
+    assert params["shut_down_mode"]["kind"] is InputKind.STRING
+    assert params["start_up_mode"]["choices"] == {None, "snow_depth"}
+    assert params["start_date"]["kind"] is InputKind.DATE
 
 
 def test_default_formatter():