+++ xclim/testing/tests/test_temperature.py
@@ -753,18 +753,12 @@ class TestDailyFreezeThaw:
         # put a nan somewhere
         tasmin.values[180, 1, 0] = np.nan
 
-        with pytest.warns(FutureWarning) as record:
-            frzthw = atmos.daily_freezethaw_cycles(tasmin, tasmax, freq="YS")
+        frzthw = atmos.daily_freezethaw_cycles(tasmin, tasmax, freq="YS")
 
         min1 = tasmin.values[:, 0, 0]
         max1 = tasmax.values[:, 0, 0]
 
         frzthw1 = ((min1 < K2C) * (max1 > K2C) * 1.0).sum()
-
-        assert (
-            "This index calculation will soon require user-specified thresholds."
-            in [str(q.message) for q in record]
-        )
 
         assert np.allclose(frzthw1, frzthw.values[0, 0, 0])
 
@@ -1215,3 +1209,27 @@ def test_degree_days_exceedance_date():
             freq="YS",
         )
         np.testing.assert_array_equal(out, np.array([[np.nan, 280, 241, 244]]).T)
+
+
+def test_warm_spell_duration_index():
+    tasmax = open_dataset("ERA5/daily_surface_cancities_1990-1993.nc").tasmax
+    tx90 = percentile_doy(tasmax, window=5, per=90)
+
+    out = atmos.warm_spell_duration_index(
+        tasmax=tasmax, tx90=tx90, window=3, freq="AS-JUL"
+    )
+    np.testing.assert_array_equal(out[0, :, 0], np.array([np.nan, 3, 0, 0, np.nan]))
+    assert (
+        "Annual total number of days within spells of at least 3 days"
+        in out.description
+    )
+
+
+def test_maximum_consecutive_warm_days():
+    tasmax = open_dataset("ERA5/daily_surface_cancities_1990-1993.nc").tasmax
+    out = atmos.maximum_consecutive_warm_days(tasmax)
+    np.testing.assert_array_equal(out[1, :], np.array([13, 21, 6, 10]))
+    assert (
+        "Annual longest spell of consecutive days with tmax above 25 degc."
+        in out.description
+    )